requirements.txt
django
pillow
gunicorn
fabric
click
google-api-python-client
google-auth-httplib2
google-auth-oauthlib